+++ account/keys.go
@@ -1,4 +1,4 @@
-package compute
+package account
 
 import (
 	"context"
